// ui/generated-api-clients/avatar/models/CourseAvatarSlot.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Service Video-Generation APIs
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const CourseAvatarSlot = {
    Default: 'default',
    Beginning: 'beginning',
    Ending: 'ending'
} as const;
export type CourseAvatarSlot = typeof CourseAvatarSlot[keyof typeof CourseAvatarSlot];


export function instanceOfCourseAvatarSlot(value: any): boolean {
    for (const key in CourseAvatarSlot) {
        if (Object.prototype.hasOwnProperty.call(CourseAvatarSlot, key)) {
            if (CourseAvatarSlot[key as keyof typeof CourseAvatarSlot] === value) {
                return true;
            }
        }
    }
    return false;
}

export function CourseAvatarSlotFromJSON(json: any): CourseAvatarSlot {
    return CourseAvatarSlotFromJSONTyped(json, false);
}

export function CourseAvatarSlotFromJSONTyped(json: any, ignoreDiscriminator: boolean): CourseAvatarSlot {
    return json as CourseAvatarSlot;
}

export function CourseAvatarSlotToJSON(value?: CourseAvatarSlot | null): any {
    return value as any;
}

export function CourseAvatarSlotToJSONTyped(value: any, ignoreDiscriminator: boolean): CourseAvatarSlot {
    return value as CourseAvatarSlot;
}